requirements.txt
fastapi>=0.68.0
fastmcp>=2.10.6
httpx>=0.26.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
python-dotenv>=1.0.0
uvicorn>=0.15.0
pydantic[all]>=2.0.0 